+++ cli/commands/create-app.ts
@@ -76,6 +76,14 @@ function generateDefinitionFile(
     url: '${appUrl}',
     groups: {},
     tags: [],
+    links: {
+      learn: '',
+      github: '',
+      twitter: '',
+      telegram: '',
+      discord: '',
+      medium: '',
+    },
     supportedNetworks: {
       ${networks.map(n => `[Network.${networkToKey[n]}]: [ProtocolAction.VIEW]`).join(',\n')}
     },