+++ src/apps/trader-joe/index.ts
@@ -1,0 +1,1 @@
+export { TRADER_JOE_DEFINITION, TraderJoeAppDefinition } from './trader-joe.definition';