+++ src/lib/index.ts
@@ -1,4 +1,5 @@
-export { AppToolkitHelperRegistry, AppToolkitHelpers } from '~app-toolkit/app-toolkit.helpers';
+// Please DO NOT add anything in this file for the time being. We have some
+// really strange transpilation issue + nest dynamic imports.
 export { AppsModule } from '~apps/apps.module';
-export { ExternalAppImport } from '~app/app.dynamic-module';
 export { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
+export { AppToolkitHelperRegistry, AppToolkitHelpers } from '~app-toolkit/app-toolkit.helpers';